secure back end

+++ pages/api/statement.js
@@ -1,27 +1,26 @@
-import { Statement } from '../../models'
-import applyMiddleware, { jwtFromReqOrCtx } from '../../util'
-import { User } from '../../models'
+import { Statement, User } from '../../models'
+import applyMiddleware from '../../util'
+import { getSession } from 'next-auth/client'
 
 export default applyMiddleware(async (req, res) => {
   try {
     const { method, body, query } = req
+    const session = await getSession({ req })
+    if (!session) throw 'Unauthorized'
     if (method === 'POST') {
-      const jwt = jwtFromReqOrCtx(req)
-      const user = await User.findOne({ email: jwt.email.toLowerCase() })
-      if (!user) throw `Could Not find ${jwt.email}`
+      const user = await User.findOne({ email: session.user.email.toLowerCase() })
+      if (!user) throw `Could Not find ${session.user.email}`
       const statement = await Statement.create({ user: user.id, alias: user.alias, description: body.description, amount: Number(body.amount)})
         .catch(err => {throw err._message})
       res.status(200).json(statement)
     } else if (method === 'GET') {
-      const jwt = jwtFromReqOrCtx(req)
-      const user = await User.findOne({ email: jwt.email.toLowerCase() })
-      if (!user) throw `Could Not find ${jwt.email}`
+      const user = await User.findOne({ email: session.user.email.toLowerCase() })
+      if (!user) throw `Could Not find ${session.user.email}`
       const agg = await getAgg()
         .catch(err => {throw err})
       res.status(200).json(agg[0])
     } else if (method === 'DELETE') {
-      const jwt = jwtFromReqOrCtx(req)
-      const user = await User.findOne({ email: jwt.email.toLowerCase() })
+      const user = await User.findOne({ email: session.user.email.toLowerCase() })
       if (!user.admin) throw 'Unathorized'
       const statement = await Statement.deleteOne({ _id: query.id })
         .catch(err => {throw err._message})
